Refine compaction scheduler
